refactor(Timeline): Rename Cypress spec and selectors
