+++ packages/react-component-library/cypress/selectors/index.ts
@@ -11,7 +11,7 @@ import rangeSlider from './RangeSlider'
 import select from './Select'
 import tabNav from './TabNav'
 import tabSet from './TabSet'
-import timeline from './timeline'
+import timeline from './Timeline'
 
 export default {
   autocomplete,
